Decouple player physics, movement, add motion provider, player manager. Add gameManager injection.
