+++ Assets/Scripts/Game/GameManagement/IGameManager.cs
@@ -3,4 +3,5 @@ using UnityEngine;
 public interface IGameManager : IService
 {
     public XRManager XR { get; }
+    public PlayerManager Player { get; }
 }
