+++ Assets/Scripts/Movement/PlayerMovement.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+
+
+public class PlayerMovement : GameBehaviour
+{
+    [SerializeField]
+    GestureMoveProvider moveProvider;
+
+    private void Start()
+    {
+        moveProvider.HandleMovementFromHandMotion();
+
+        gameManager.Player.Physics.OnUpdateMomentum += moveProvider.HandleMovementFromHandMotion;
+    }
+}
