+++ Assets/Scripts/Services/Delegates.cs
@@ -1,0 +1,7 @@
+
+using UnityEngine;
+
+public class Delegates : MonoBehaviour
+{
+    public delegate void UpdateAction();
+}
