+++ Assets/Scripts/Services/Delegates.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 45e6f3f04dab1654aa017023dabba95f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
